feat : ocrResult 좌표 타입 int형에서 double형으로 변경

See merge request s13-webmobile4-sub1/S13P11D204!102

+++ bookglebookgle-server/src/main/java/com/example/bookglebookgleserver/ocr/dto/OcrTextBlockDto.java
@@ -10,8 +10,8 @@ import lombok.*;
 public class OcrTextBlockDto {
     private int pageNumber;
     private String text;
-    private int rectX;
-    private int rectY;
-    private int rectW;
-    private int rectH;
+    private double rectX;
+    private double rectY;
+    private double rectW;
+    private double rectH;
 }
